# docs/api-documentation-openapi.yaml
openapi: 3.0.0
...
info:
  title: OpenID4VP Verifier API - Inji Verify
  version: 1.0.2
  contact:
    name: MOSIP
    url: "https://mosip.io"
    email: mosip.emailnotifier@gmail.com
  license:
    name: MPL-2.0
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  description: API for OpenID4VP verifier supporting direct_post and cross-device flow
paths:
  /vp-request:
    post:
      summary: Create a new verification request
      description: "Using its response the UI can build /authorize endpoint and render it in qrcode form based on cross or same device flow. example - openid4vp://authorize?{RESPONSE}"
      operationId: createVPRequest
      tags:
        - VP Request Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VPRequestCreate"
            example:
              clientId: client123
              presentationDefinitionId: age_verification
      responses:
        "201":
          description: VP Request created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VPRequestResponse"
              example:
                transactionId: txn_12345
                requestId: req_12345
                authorizationDetails: {}
                expiresAt: 123456789876
        "400":
          description: VP Request not created due to bad request. Fields missing or type missmatch in body sent.
        "404":
          description: VP Request not created due to PresentationDefiniton not found for given ID.
      x-stoplight:
        id: xs9xoxyyqg5qj
  "/vp-request/{requestId}/status":
    get:
      summary: Get VP Request status
      description: This API returns the current status of the transaction and returns not found if there is no transaction with given request id.
      operationId: getVPRequestStatus
      tags:
        - VP Request Management
      parameters:
        - name: requestId
          description: Request Id for which you want to get the status for
          in: path
          required: true
          schema:
            type: string
          example: req_12345
      responses:
        "200":
          description: VP Request status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/status"
        "404":
          description: VP Request not found for given request id
      x-stoplight:
        id: fnqvm2pbdv6wy
...
components:
  schemas:
    VPRequestCreate:
      type: object
      required:
        - clientId
      allOf:
        - oneOf:
            - required:
                - presentationDefinitionId
            - required:
                - presentationDefinition
      properties:
        clientId:
          type: string
          description: ID of the client asking to create a new Verifiable Presentation request
        presentationDefinitionId:
          type: string
          description: Either presentationDefinitionId or presentationDefinition should be passed. A predefined presentation definition Id so that the user does not need to pass the presentation definition
        presentationDefinition:
          type: string
          description: Either presentationDefinitionId or presentationDefinition should be passed. Verifiable Presentation definition for getting the credential
        transactionId:
          description: A unique id of the verifiable presentation request transaction. This is optional and will be generated on server if not passed.
          type: string
      x-stoplight:
        id: ayf2p7ofmj2g1
      title: ""
    VPRequestResponse:
      type: object
      properties:
        transactionId:
          description: A unique id of the verifiable presentation request transaction
          type: string
        requestId:
          type: string
          description: A unique id of the verifiable presentation request
        authorizationDetails:
          type: object
          description: The authorization request object with entire presentation_definition
          $ref: "#/components/schemas/AuthorizationRequestResponse"
        expiresAt:
          type: integer
          description: Expiry of a transaction. Defaults to 5 minutes from creation of auth request.
      x-stoplight:
        id: wkbn3jolevhw7
    status:
      type: object
      description: Status of the verifiable presentation request transaction
      properties:
        status:
          type: string
          description: current status
          enum:
            - ACTIVE
            - VP_SUBMITTED
            - EXPIRED
...
    AuthorizationRequestResponse:
      title: AuthorizationRequestResponse
      x-stoplight:
        id: qiiayk4b5xbsu
      type: object
      properties:
        responseUri:
          type: string
          description: URL for the wallet to submit the verifiable presentation
          format: uri
        responseType:
          type: string
          description: Response type of the verifiable presentation
        clientId:
          type: string
          description: Client id of the verifier
        nonce:
          type: string
          description: Nonce for encryption. It is used to securely bind the Verifiable Presentation(s) provided by the Wallet to the particular transaction
        presentationDefinitionUri:
          type: string
          format: uri
          description: A string containing a Presentation Definition JSON object
        iat:
          type: integer
          format: unixtime
          description: Issued time
...
  securitySchemes:
    ClientIdSecret:
      type: http
      scheme: basic